Changed default behaviour of Path(), changed modules where necessary
git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@757 9dfea958-5e0b-0410-aa7e-b09452dc4efc

+++ lib/models/repository.py
@@ -143,6 +143,9 @@ class Repository(object):
 		config = Config()
 		reposdir = config.get('svn', 'repositories')
 		newrepos = Path(reposdir) + self.name
+		if not newrepos.absolute:
+			raise Exception("Error, repository path is relative, this should be fixed")
+
 		cmd = 'rm -rf "%s"' % newrepos
 		(exitstatus, outtext) = commands.getstatusoutput(cmd)
 		if exitstatus == 0:
